Unfinished reversed extra credit
Here I began to do the extra credit lab. I don't expect many or maybe any points to be earned but I was trying for a few hours in an attempt to get more practice. I would like some credit but I understand that this is not my best work and that it will through a segmentation error as I could never get head to work like I wanted it to. I tried. I also apologize for the lack of comments the end of the semester has been getting hectic.